ffmpeg_sys_next

Enum AVCodecID

#[repr(u32)]
pub enum AVCodecID {

    AV_CODEC_ID_NONE = 0,
    AV_CODEC_ID_MPEG1VIDEO = 1,
    AV_CODEC_ID_MPEG2VIDEO = 2,
    AV_CODEC_ID_H261 = 3,
    AV_CODEC_ID_H263 = 4,
    AV_CODEC_ID_RV10 = 5,
    AV_CODEC_ID_RV20 = 6,
    AV_CODEC_ID_MJPEG = 7,
    AV_CODEC_ID_MJPEGB = 8,
    AV_CODEC_ID_LJPEG = 9,
    AV_CODEC_ID_SP5X = 10,
    AV_CODEC_ID_JPEGLS = 11,
    AV_CODEC_ID_MPEG4 = 12,
    AV_CODEC_ID_RAWVIDEO = 13,
    AV_CODEC_ID_MSMPEG4V1 = 14,
    AV_CODEC_ID_MSMPEG4V2 = 15,
    AV_CODEC_ID_MSMPEG4V3 = 16,
    AV_CODEC_ID_WMV1 = 17,
    AV_CODEC_ID_WMV2 = 18,
    AV_CODEC_ID_H263P = 19,
    AV_CODEC_ID_H263I = 20,
    AV_CODEC_ID_FLV1 = 21,
    AV_CODEC_ID_SVQ1 = 22,
    AV_CODEC_ID_SVQ3 = 23,
    AV_CODEC_ID_DVVIDEO = 24,
    AV_CODEC_ID_HUFFYUV = 25,
    AV_CODEC_ID_CYUV = 26,
    AV_CODEC_ID_H264 = 27,
    AV_CODEC_ID_INDEO3 = 28,
    AV_CODEC_ID_VP3 = 29,
    AV_CODEC_ID_THEORA = 30,
    AV_CODEC_ID_ASV1 = 31,
    AV_CODEC_ID_ASV2 = 32,
    AV_CODEC_ID_FFV1 = 33,
    AV_CODEC_ID_4XM = 34,
    AV_CODEC_ID_VCR1 = 35,
    AV_CODEC_ID_CLJR = 36,
    AV_CODEC_ID_MDEC = 37,
    AV_CODEC_ID_ROQ = 38,
    AV_CODEC_ID_INTERPLAY_VIDEO = 39,
    AV_CODEC_ID_XAN_WC3 = 40,
    AV_CODEC_ID_XAN_WC4 = 41,
    AV_CODEC_ID_RPZA = 42,
    AV_CODEC_ID_CINEPAK = 43,
    AV_CODEC_ID_WS_VQA = 44,
    AV_CODEC_ID_MSRLE = 45,
    AV_CODEC_ID_MSVIDEO1 = 46,
    AV_CODEC_ID_IDCIN = 47,
    AV_CODEC_ID_8BPS = 48,
    AV_CODEC_ID_SMC = 49,
    AV_CODEC_ID_FLIC = 50,
    AV_CODEC_ID_TRUEMOTION1 = 51,
    AV_CODEC_ID_VMDVIDEO = 52,
    AV_CODEC_ID_MSZH = 53,
    AV_CODEC_ID_ZLIB = 54,
    AV_CODEC_ID_QTRLE = 55,
    AV_CODEC_ID_TSCC = 56,
    AV_CODEC_ID_ULTI = 57,
    AV_CODEC_ID_QDRAW = 58,
    AV_CODEC_ID_VIXL = 59,
    AV_CODEC_ID_QPEG = 60,
    AV_CODEC_ID_PNG = 61,
    AV_CODEC_ID_PPM = 62,
    AV_CODEC_ID_PBM = 63,
    AV_CODEC_ID_PGM = 64,
    AV_CODEC_ID_PGMYUV = 65,
    AV_CODEC_ID_PAM = 66,
    AV_CODEC_ID_FFVHUFF = 67,
    AV_CODEC_ID_RV30 = 68,
    AV_CODEC_ID_RV40 = 69,
    AV_CODEC_ID_VC1 = 70,
    AV_CODEC_ID_WMV3 = 71,
    AV_CODEC_ID_LOCO = 72,
    AV_CODEC_ID_WNV1 = 73,
    AV_CODEC_ID_AASC = 74,
    AV_CODEC_ID_INDEO2 = 75,
    AV_CODEC_ID_FRAPS = 76,
    AV_CODEC_ID_TRUEMOTION2 = 77,
    AV_CODEC_ID_BMP = 78,
    AV_CODEC_ID_CSCD = 79,
    AV_CODEC_ID_MMVIDEO = 80,
    AV_CODEC_ID_ZMBV = 81,
    AV_CODEC_ID_AVS = 82,
    AV_CODEC_ID_SMACKVIDEO = 83,
    AV_CODEC_ID_NUV = 84,
    AV_CODEC_ID_KMVC = 85,
    AV_CODEC_ID_FLASHSV = 86,
    AV_CODEC_ID_CAVS = 87,
    AV_CODEC_ID_JPEG2000 = 88,
    AV_CODEC_ID_VMNC = 89,
    AV_CODEC_ID_VP5 = 90,
    AV_CODEC_ID_VP6 = 91,
    AV_CODEC_ID_VP6F = 92,
    AV_CODEC_ID_TARGA = 93,
    AV_CODEC_ID_DSICINVIDEO = 94,
    AV_CODEC_ID_TIERTEXSEQVIDEO = 95,
    AV_CODEC_ID_TIFF = 96,
    AV_CODEC_ID_GIF = 97,
    AV_CODEC_ID_DXA = 98,
    AV_CODEC_ID_DNXHD = 99,
    AV_CODEC_ID_THP = 100,
    AV_CODEC_ID_SGI = 101,
    AV_CODEC_ID_C93 = 102,
    AV_CODEC_ID_BETHSOFTVID = 103,
    AV_CODEC_ID_PTX = 104,
    AV_CODEC_ID_TXD = 105,
    AV_CODEC_ID_VP6A = 106,
    AV_CODEC_ID_AMV = 107,
    AV_CODEC_ID_VB = 108,
    AV_CODEC_ID_PCX = 109,
    AV_CODEC_ID_SUNRAST = 110,
    AV_CODEC_ID_INDEO4 = 111,
    AV_CODEC_ID_INDEO5 = 112,
    AV_CODEC_ID_MIMIC = 113,
    AV_CODEC_ID_RL2 = 114,
    AV_CODEC_ID_ESCAPE124 = 115,
    AV_CODEC_ID_DIRAC = 116,
    AV_CODEC_ID_BFI = 117,
    AV_CODEC_ID_CMV = 118,
    AV_CODEC_ID_MOTIONPIXELS = 119,
    AV_CODEC_ID_TGV = 120,
    AV_CODEC_ID_TGQ = 121,
    AV_CODEC_ID_TQI = 122,
    AV_CODEC_ID_AURA = 123,
    AV_CODEC_ID_AURA2 = 124,
    AV_CODEC_ID_V210X = 125,
    AV_CODEC_ID_TMV = 126,
    AV_CODEC_ID_V210 = 127,
    AV_CODEC_ID_DPX = 128,
    AV_CODEC_ID_MAD = 129,
    AV_CODEC_ID_FRWU = 130,
    AV_CODEC_ID_FLASHSV2 = 131,
    AV_CODEC_ID_CDGRAPHICS = 132,
    AV_CODEC_ID_R210 = 133,
    AV_CODEC_ID_ANM = 134,
    AV_CODEC_ID_BINKVIDEO = 135,
    AV_CODEC_ID_IFF_ILBM = 136,
    AV_CODEC_ID_KGV1 = 137,
    AV_CODEC_ID_YOP = 138,
    AV_CODEC_ID_VP8 = 139,
    AV_CODEC_ID_PICTOR = 140,
    AV_CODEC_ID_ANSI = 141,
    AV_CODEC_ID_A64_MULTI = 142,
    AV_CODEC_ID_A64_MULTI5 = 143,
    AV_CODEC_ID_R10K = 144,
    AV_CODEC_ID_MXPEG = 145,
    AV_CODEC_ID_LAGARITH = 146,
    AV_CODEC_ID_PRORES = 147,
    AV_CODEC_ID_JV = 148,
    AV_CODEC_ID_DFA = 149,
    AV_CODEC_ID_WMV3IMAGE = 150,
    AV_CODEC_ID_VC1IMAGE = 151,
    AV_CODEC_ID_UTVIDEO = 152,
    AV_CODEC_ID_BMV_VIDEO = 153,
    AV_CODEC_ID_VBLE = 154,
    AV_CODEC_ID_DXTORY = 155,
    AV_CODEC_ID_V410 = 156,
    AV_CODEC_ID_XWD = 157,
    AV_CODEC_ID_CDXL = 158,
    AV_CODEC_ID_XBM = 159,
    AV_CODEC_ID_ZEROCODEC = 160,
    AV_CODEC_ID_MSS1 = 161,
    AV_CODEC_ID_MSA1 = 162,
    AV_CODEC_ID_TSCC2 = 163,
    AV_CODEC_ID_MTS2 = 164,
    AV_CODEC_ID_CLLC = 165,
    AV_CODEC_ID_MSS2 = 166,
    AV_CODEC_ID_VP9 = 167,
    AV_CODEC_ID_AIC = 168,
    AV_CODEC_ID_ESCAPE130 = 169,
    AV_CODEC_ID_G2M = 170,
    AV_CODEC_ID_WEBP = 171,
    AV_CODEC_ID_HNM4_VIDEO = 172,
    AV_CODEC_ID_HEVC = 173,
    AV_CODEC_ID_FIC = 174,
    AV_CODEC_ID_ALIAS_PIX = 175,
    AV_CODEC_ID_BRENDER_PIX = 176,
    AV_CODEC_ID_PAF_VIDEO = 177,
    AV_CODEC_ID_EXR = 178,
    AV_CODEC_ID_VP7 = 179,
    AV_CODEC_ID_SANM = 180,
    AV_CODEC_ID_SGIRLE = 181,
    AV_CODEC_ID_MVC1 = 182,
    AV_CODEC_ID_MVC2 = 183,
    AV_CODEC_ID_HQX = 184,
    AV_CODEC_ID_TDSC = 185,
    AV_CODEC_ID_HQ_HQA = 186,
    AV_CODEC_ID_HAP = 187,
    AV_CODEC_ID_DDS = 188,
    AV_CODEC_ID_DXV = 189,
    AV_CODEC_ID_SCREENPRESSO = 190,
    AV_CODEC_ID_RSCC = 191,
    AV_CODEC_ID_AVS2 = 192,
    AV_CODEC_ID_PGX = 193,
    AV_CODEC_ID_AVS3 = 194,
    AV_CODEC_ID_MSP2 = 195,
    AV_CODEC_ID_VVC = 196,
    AV_CODEC_ID_Y41P = 197,
    AV_CODEC_ID_AVRP = 198,
    AV_CODEC_ID_012V = 199,
    AV_CODEC_ID_AVUI = 200,
    AV_CODEC_ID_TARGA_Y216 = 201,
    AV_CODEC_ID_V308 = 202,
    AV_CODEC_ID_V408 = 203,
    AV_CODEC_ID_YUV4 = 204,
    AV_CODEC_ID_AVRN = 205,
    AV_CODEC_ID_CPIA = 206,
    AV_CODEC_ID_XFACE = 207,
    AV_CODEC_ID_SNOW = 208,
    AV_CODEC_ID_SMVJPEG = 209,
    AV_CODEC_ID_APNG = 210,
    AV_CODEC_ID_DAALA = 211,
    AV_CODEC_ID_CFHD = 212,
    AV_CODEC_ID_TRUEMOTION2RT = 213,
    AV_CODEC_ID_M101 = 214,
    AV_CODEC_ID_MAGICYUV = 215,
    AV_CODEC_ID_SHEERVIDEO = 216,
    AV_CODEC_ID_YLC = 217,
    AV_CODEC_ID_PSD = 218,
    AV_CODEC_ID_PIXLET = 219,
    AV_CODEC_ID_SPEEDHQ = 220,
    AV_CODEC_ID_FMVC = 221,
    AV_CODEC_ID_SCPR = 222,
    AV_CODEC_ID_CLEARVIDEO = 223,
    AV_CODEC_ID_XPM = 224,
    AV_CODEC_ID_AV1 = 225,
    AV_CODEC_ID_BITPACKED = 226,
    AV_CODEC_ID_MSCC = 227,
    AV_CODEC_ID_SRGC = 228,
    AV_CODEC_ID_SVG = 229,
    AV_CODEC_ID_GDV = 230,
    AV_CODEC_ID_FITS = 231,
    AV_CODEC_ID_IMM4 = 232,
    AV_CODEC_ID_PROSUMER = 233,
    AV_CODEC_ID_MWSC = 234,
    AV_CODEC_ID_WCMV = 235,
    AV_CODEC_ID_RASC = 236,
    AV_CODEC_ID_HYMT = 237,
    AV_CODEC_ID_ARBC = 238,
    AV_CODEC_ID_AGM = 239,
    AV_CODEC_ID_LSCR = 240,
    AV_CODEC_ID_VP4 = 241,
    AV_CODEC_ID_IMM5 = 242,
    AV_CODEC_ID_MVDV = 243,
    AV_CODEC_ID_MVHA = 244,
    AV_CODEC_ID_CDTOONS = 245,
    AV_CODEC_ID_MV30 = 246,
    AV_CODEC_ID_NOTCHLC = 247,
    AV_CODEC_ID_PFM = 248,
    AV_CODEC_ID_MOBICLIP = 249,
    AV_CODEC_ID_PHOTOCD = 250,
    AV_CODEC_ID_IPU = 251,
    AV_CODEC_ID_ARGO = 252,
    AV_CODEC_ID_CRI = 253,
    AV_CODEC_ID_SIMBIOSIS_IMX = 254,
    AV_CODEC_ID_SGA_VIDEO = 255,
    AV_CODEC_ID_GEM = 256,
    AV_CODEC_ID_VBN = 257,
    AV_CODEC_ID_JPEGXL = 258,
    AV_CODEC_ID_QOI = 259,
    AV_CODEC_ID_PHM = 260,
    AV_CODEC_ID_RADIANCE_HDR = 261,
    AV_CODEC_ID_WBMP = 262,
    AV_CODEC_ID_MEDIA100 = 263,
    AV_CODEC_ID_VQC = 264,
    AV_CODEC_ID_PDV = 265,
    AV_CODEC_ID_EVC = 266,
    AV_CODEC_ID_RTV1 = 267,
    AV_CODEC_ID_VMIX = 268,
    AV_CODEC_ID_LEAD = 269,
    AV_CODEC_ID_PCM_S16LE = 65_536,
    AV_CODEC_ID_PCM_S16BE = 65_537,
    AV_CODEC_ID_PCM_U16LE = 65_538,
    AV_CODEC_ID_PCM_U16BE = 65_539,
    AV_CODEC_ID_PCM_S8 = 65_540,
    AV_CODEC_ID_PCM_U8 = 65_541,
    AV_CODEC_ID_PCM_MULAW = 65_542,
    AV_CODEC_ID_PCM_ALAW = 65_543,
    AV_CODEC_ID_PCM_S32LE = 65_544,
    AV_CODEC_ID_PCM_S32BE = 65_545,
    AV_CODEC_ID_PCM_U32LE = 65_546,
    AV_CODEC_ID_PCM_U32BE = 65_547,
    AV_CODEC_ID_PCM_S24LE = 65_548,
    AV_CODEC_ID_PCM_S24BE = 65_549,
    AV_CODEC_ID_PCM_U24LE = 65_550,
    AV_CODEC_ID_PCM_U24BE = 65_551,
    AV_CODEC_ID_PCM_S24DAUD = 65_552,
    AV_CODEC_ID_PCM_ZORK = 65_553,
    AV_CODEC_ID_PCM_S16LE_PLANAR = 65_554,
    AV_CODEC_ID_PCM_DVD = 65_555,
    AV_CODEC_ID_PCM_F32BE = 65_556,
    AV_CODEC_ID_PCM_F32LE = 65_557,
    AV_CODEC_ID_PCM_F64BE = 65_558,
    AV_CODEC_ID_PCM_F64LE = 65_559,
    AV_CODEC_ID_PCM_BLURAY = 65_560,
    AV_CODEC_ID_PCM_LXF = 65_561,
    AV_CODEC_ID_S302M = 65_562,
    AV_CODEC_ID_PCM_S8_PLANAR = 65_563,
    AV_CODEC_ID_PCM_S24LE_PLANAR = 65_564,
    AV_CODEC_ID_PCM_S32LE_PLANAR = 65_565,
    AV_CODEC_ID_PCM_S16BE_PLANAR = 65_566,
    AV_CODEC_ID_PCM_S64LE = 65_567,
    AV_CODEC_ID_PCM_S64BE = 65_568,
    AV_CODEC_ID_PCM_F16LE = 65_569,
    AV_CODEC_ID_PCM_F24LE = 65_570,
    AV_CODEC_ID_PCM_VIDC = 65_571,
    AV_CODEC_ID_PCM_SGA = 65_572,
    AV_CODEC_ID_ADPCM_IMA_QT = 69_632,
    AV_CODEC_ID_ADPCM_IMA_WAV = 69_633,
    AV_CODEC_ID_ADPCM_IMA_DK3 = 69_634,
    AV_CODEC_ID_ADPCM_IMA_DK4 = 69_635,
    AV_CODEC_ID_ADPCM_IMA_WS = 69_636,
    AV_CODEC_ID_ADPCM_IMA_SMJPEG = 69_637,
    AV_CODEC_ID_ADPCM_MS = 69_638,
    AV_CODEC_ID_ADPCM_4XM = 69_639,
    AV_CODEC_ID_ADPCM_XA = 69_640,
    AV_CODEC_ID_ADPCM_ADX = 69_641,
    AV_CODEC_ID_ADPCM_EA = 69_642,
    AV_CODEC_ID_ADPCM_G726 = 69_643,
    AV_CODEC_ID_ADPCM_CT = 69_644,
    AV_CODEC_ID_ADPCM_SWF = 69_645,
    AV_CODEC_ID_ADPCM_YAMAHA = 69_646,
    AV_CODEC_ID_ADPCM_SBPRO_4 = 69_647,
    AV_CODEC_ID_ADPCM_SBPRO_3 = 69_648,
    AV_CODEC_ID_ADPCM_SBPRO_2 = 69_649,
    AV_CODEC_ID_ADPCM_THP = 69_650,
    AV_CODEC_ID_ADPCM_IMA_AMV = 69_651,
    AV_CODEC_ID_ADPCM_EA_R1 = 69_652,
    AV_CODEC_ID_ADPCM_EA_R3 = 69_653,
    AV_CODEC_ID_ADPCM_EA_R2 = 69_654,
    AV_CODEC_ID_ADPCM_IMA_EA_SEAD = 69_655,
    AV_CODEC_ID_ADPCM_IMA_EA_EACS = 69_656,
    AV_CODEC_ID_ADPCM_EA_XAS = 69_657,
    AV_CODEC_ID_ADPCM_EA_MAXIS_XA = 69_658,
    AV_CODEC_ID_ADPCM_IMA_ISS = 69_659,
    AV_CODEC_ID_ADPCM_G722 = 69_660,
    AV_CODEC_ID_ADPCM_IMA_APC = 69_661,
    AV_CODEC_ID_ADPCM_VIMA = 69_662,
    AV_CODEC_ID_ADPCM_AFC = 69_663,
    AV_CODEC_ID_ADPCM_IMA_OKI = 69_664,
    AV_CODEC_ID_ADPCM_DTK = 69_665,
    AV_CODEC_ID_ADPCM_IMA_RAD = 69_666,
    AV_CODEC_ID_ADPCM_G726LE = 69_667,
    AV_CODEC_ID_ADPCM_THP_LE = 69_668,
    AV_CODEC_ID_ADPCM_PSX = 69_669,
    AV_CODEC_ID_ADPCM_AICA = 69_670,
    AV_CODEC_ID_ADPCM_IMA_DAT4 = 69_671,
    AV_CODEC_ID_ADPCM_MTAF = 69_672,
    AV_CODEC_ID_ADPCM_AGM = 69_673,
    AV_CODEC_ID_ADPCM_ARGO = 69_674,
    AV_CODEC_ID_ADPCM_IMA_SSI = 69_675,
    AV_CODEC_ID_ADPCM_ZORK = 69_676,
    AV_CODEC_ID_ADPCM_IMA_APM = 69_677,
    AV_CODEC_ID_ADPCM_IMA_ALP = 69_678,
    AV_CODEC_ID_ADPCM_IMA_MTF = 69_679,
    AV_CODEC_ID_ADPCM_IMA_CUNNING = 69_680,
    AV_CODEC_ID_ADPCM_IMA_MOFLEX = 69_681,
    AV_CODEC_ID_ADPCM_IMA_ACORN = 69_682,
    AV_CODEC_ID_ADPCM_XMD = 69_683,
    AV_CODEC_ID_AMR_NB = 73_728,
    AV_CODEC_ID_AMR_WB = 73_729,
    AV_CODEC_ID_RA_144 = 77_824,
    AV_CODEC_ID_RA_288 = 77_825,
    AV_CODEC_ID_ROQ_DPCM = 81_920,
    AV_CODEC_ID_INTERPLAY_DPCM = 81_921,
    AV_CODEC_ID_XAN_DPCM = 81_922,
    AV_CODEC_ID_SOL_DPCM = 81_923,
    AV_CODEC_ID_SDX2_DPCM = 81_924,
    AV_CODEC_ID_GREMLIN_DPCM = 81_925,
    AV_CODEC_ID_DERF_DPCM = 81_926,
    AV_CODEC_ID_WADY_DPCM = 81_927,
    AV_CODEC_ID_CBD2_DPCM = 81_928,
    AV_CODEC_ID_MP2 = 86_016,
    AV_CODEC_ID_MP3 = 86_017,
    AV_CODEC_ID_AAC = 86_018,
    AV_CODEC_ID_AC3 = 86_019,
    AV_CODEC_ID_DTS = 86_020,
    AV_CODEC_ID_VORBIS = 86_021,
    AV_CODEC_ID_DVAUDIO = 86_022,
    AV_CODEC_ID_WMAV1 = 86_023,
    AV_CODEC_ID_WMAV2 = 86_024,
    AV_CODEC_ID_MACE3 = 86_025,
    AV_CODEC_ID_MACE6 = 86_026,
    AV_CODEC_ID_VMDAUDIO = 86_027,
    AV_CODEC_ID_FLAC = 86_028,
    AV_CODEC_ID_MP3ADU = 86_029,
    AV_CODEC_ID_MP3ON4 = 86_030,
    AV_CODEC_ID_SHORTEN = 86_031,
    AV_CODEC_ID_ALAC = 86_032,
    AV_CODEC_ID_WESTWOOD_SND1 = 86_033,
    AV_CODEC_ID_GSM = 86_034,
    AV_CODEC_ID_QDM2 = 86_035,
    AV_CODEC_ID_COOK = 86_036,
    AV_CODEC_ID_TRUESPEECH = 86_037,
    AV_CODEC_ID_TTA = 86_038,
    AV_CODEC_ID_SMACKAUDIO = 86_039,
    AV_CODEC_ID_QCELP = 86_040,
    AV_CODEC_ID_WAVPACK = 86_041,
    AV_CODEC_ID_DSICINAUDIO = 86_042,
    AV_CODEC_ID_IMC = 86_043,
    AV_CODEC_ID_MUSEPACK7 = 86_044,
    AV_CODEC_ID_MLP = 86_045,
    AV_CODEC_ID_GSM_MS = 86_046,
    AV_CODEC_ID_ATRAC3 = 86_047,
    AV_CODEC_ID_APE = 86_048,
    AV_CODEC_ID_NELLYMOSER = 86_049,
    AV_CODEC_ID_MUSEPACK8 = 86_050,
    AV_CODEC_ID_SPEEX = 86_051,
    AV_CODEC_ID_WMAVOICE = 86_052,
    AV_CODEC_ID_WMAPRO = 86_053,
    AV_CODEC_ID_WMALOSSLESS = 86_054,
    AV_CODEC_ID_ATRAC3P = 86_055,
    AV_CODEC_ID_EAC3 = 86_056,
    AV_CODEC_ID_SIPR = 86_057,
    AV_CODEC_ID_MP1 = 86_058,
    AV_CODEC_ID_TWINVQ = 86_059,
    AV_CODEC_ID_TRUEHD = 86_060,
    AV_CODEC_ID_MP4ALS = 86_061,
    AV_CODEC_ID_ATRAC1 = 86_062,
    AV_CODEC_ID_BINKAUDIO_RDFT = 86_063,
    AV_CODEC_ID_BINKAUDIO_DCT = 86_064,
    AV_CODEC_ID_AAC_LATM = 86_065,
    AV_CODEC_ID_QDMC = 86_066,
    AV_CODEC_ID_CELT = 86_067,
    AV_CODEC_ID_G723_1 = 86_068,
    AV_CODEC_ID_G729 = 86_069,
    AV_CODEC_ID_8SVX_EXP = 86_070,
    AV_CODEC_ID_8SVX_FIB = 86_071,
    AV_CODEC_ID_BMV_AUDIO = 86_072,
    AV_CODEC_ID_RALF = 86_073,
    AV_CODEC_ID_IAC = 86_074,
    AV_CODEC_ID_ILBC = 86_075,
    AV_CODEC_ID_OPUS = 86_076,
    AV_CODEC_ID_COMFORT_NOISE = 86_077,
    AV_CODEC_ID_TAK = 86_078,
    AV_CODEC_ID_METASOUND = 86_079,
    AV_CODEC_ID_PAF_AUDIO = 86_080,
    AV_CODEC_ID_ON2AVC = 86_081,
    AV_CODEC_ID_DSS_SP = 86_082,
    AV_CODEC_ID_CODEC2 = 86_083,
    AV_CODEC_ID_FFWAVESYNTH = 86_084,
    AV_CODEC_ID_SONIC = 86_085,
    AV_CODEC_ID_SONIC_LS = 86_086,
    AV_CODEC_ID_EVRC = 86_087,
    AV_CODEC_ID_SMV = 86_088,
    AV_CODEC_ID_DSD_LSBF = 86_089,
    AV_CODEC_ID_DSD_MSBF = 86_090,
    AV_CODEC_ID_DSD_LSBF_PLANAR = 86_091,
    AV_CODEC_ID_DSD_MSBF_PLANAR = 86_092,
    AV_CODEC_ID_4GV = 86_093,
    AV_CODEC_ID_INTERPLAY_ACM = 86_094,
    AV_CODEC_ID_XMA1 = 86_095,
    AV_CODEC_ID_XMA2 = 86_096,
    AV_CODEC_ID_DST = 86_097,
    AV_CODEC_ID_ATRAC3AL = 86_098,
    AV_CODEC_ID_ATRAC3PAL = 86_099,
    AV_CODEC_ID_DOLBY_E = 86_100,
    AV_CODEC_ID_APTX = 86_101,
    AV_CODEC_ID_APTX_HD = 86_102,
    AV_CODEC_ID_SBC = 86_103,
    AV_CODEC_ID_ATRAC9 = 86_104,
    AV_CODEC_ID_HCOM = 86_105,
    AV_CODEC_ID_ACELP_KELVIN = 86_106,
    AV_CODEC_ID_MPEGH_3D_AUDIO = 86_107,
    AV_CODEC_ID_SIREN = 86_108,
    AV_CODEC_ID_HCA = 86_109,
    AV_CODEC_ID_FASTAUDIO = 86_110,
    AV_CODEC_ID_MSNSIREN = 86_111,
    AV_CODEC_ID_DFPWM = 86_112,
    AV_CODEC_ID_BONK = 86_113,
    AV_CODEC_ID_MISC4 = 86_114,
    AV_CODEC_ID_APAC = 86_115,
    AV_CODEC_ID_FTR = 86_116,
    AV_CODEC_ID_WAVARC = 86_117,
    AV_CODEC_ID_RKA = 86_118,
    AV_CODEC_ID_AC4 = 86_119,
    AV_CODEC_ID_OSQ = 86_120,
    AV_CODEC_ID_QOA = 86_121,
    AV_CODEC_ID_LC3 = 86_122,
    AV_CODEC_ID_DVD_SUBTITLE = 94_208,
    AV_CODEC_ID_DVB_SUBTITLE = 94_209,
    AV_CODEC_ID_TEXT = 94_210,
    AV_CODEC_ID_XSUB = 94_211,
    AV_CODEC_ID_SSA = 94_212,
    AV_CODEC_ID_MOV_TEXT = 94_213,
    AV_CODEC_ID_HDMV_PGS_SUBTITLE = 94_214,
    AV_CODEC_ID_DVB_TELETEXT = 94_215,
    AV_CODEC_ID_SRT = 94_216,
    AV_CODEC_ID_MICRODVD = 94_217,
    AV_CODEC_ID_EIA_608 = 94_218,
    AV_CODEC_ID_JACOSUB = 94_219,
    AV_CODEC_ID_SAMI = 94_220,
    AV_CODEC_ID_REALTEXT = 94_221,
    AV_CODEC_ID_STL = 94_222,
    AV_CODEC_ID_SUBVIEWER1 = 94_223,
    AV_CODEC_ID_SUBVIEWER = 94_224,
    AV_CODEC_ID_SUBRIP = 94_225,
    AV_CODEC_ID_WEBVTT = 94_226,
    AV_CODEC_ID_MPL2 = 94_227,
    AV_CODEC_ID_VPLAYER = 94_228,
    AV_CODEC_ID_PJS = 94_229,
    AV_CODEC_ID_ASS = 94_230,
    AV_CODEC_ID_HDMV_TEXT_SUBTITLE = 94_231,
    AV_CODEC_ID_TTML = 94_232,
    AV_CODEC_ID_ARIB_CAPTION = 94_233,
    AV_CODEC_ID_TTF = 98_304,
    AV_CODEC_ID_SCTE_35 = 98_305,
    AV_CODEC_ID_EPG = 98_306,
    AV_CODEC_ID_BINTEXT = 98_307,
    AV_CODEC_ID_XBIN = 98_308,
    AV_CODEC_ID_IDF = 98_309,
    AV_CODEC_ID_OTF = 98_310,
    AV_CODEC_ID_SMPTE_KLV = 98_311,
    AV_CODEC_ID_DVD_NAV = 98_312,
    AV_CODEC_ID_TIMED_ID3 = 98_313,
    AV_CODEC_ID_BIN_DATA = 98_314,
    AV_CODEC_ID_SMPTE_2038 = 98_315,
    AV_CODEC_ID_LCEVC = 98_316,
    AV_CODEC_ID_PROBE = 102_400,
    AV_CODEC_ID_MPEG2TS = 131_072,
    AV_CODEC_ID_MPEG4SYSTEMS = 131_073,
    AV_CODEC_ID_FFMETADATA = 135_168,
    AV_CODEC_ID_WRAPPED_AVFRAME = 135_169,
    AV_CODEC_ID_VNULL = 135_170,
    AV_CODEC_ID_ANULL = 135_171,

}

Identify the syntax and semantics of the bitstream. The principle is roughly: Two decoders with the same ID can decode the same streams. Two encoders with the same ID can encode compatible streams. There may be slight deviations from the principle due to implementation details.

If you add a codec ID to this list, add it so that

no value of an existing codec ID changes (that would break ABI),

it is as close as possible to similar codecs

After adding new codec IDs, do not forget to add an entry to the codec descriptor list and bump libavcodec minor version.

Variants

AV_CODEC_ID_NONE = 0

AV_CODEC_ID_MPEG1VIDEO = 1

AV_CODEC_ID_MPEG2VIDEO = 2

< preferred ID for MPEG-1/2 video decoding

AV_CODEC_ID_H261 = 3

AV_CODEC_ID_H263 = 4

AV_CODEC_ID_RV10 = 5

AV_CODEC_ID_RV20 = 6

AV_CODEC_ID_MJPEG = 7

AV_CODEC_ID_MJPEGB = 8

AV_CODEC_ID_LJPEG = 9

AV_CODEC_ID_SP5X = 10

AV_CODEC_ID_JPEGLS = 11

AV_CODEC_ID_MPEG4 = 12

AV_CODEC_ID_RAWVIDEO = 13

AV_CODEC_ID_MSMPEG4V1 = 14

AV_CODEC_ID_MSMPEG4V2 = 15

AV_CODEC_ID_MSMPEG4V3 = 16

AV_CODEC_ID_WMV1 = 17

AV_CODEC_ID_WMV2 = 18

AV_CODEC_ID_H263P = 19

AV_CODEC_ID_H263I = 20

AV_CODEC_ID_FLV1 = 21

AV_CODEC_ID_SVQ1 = 22

AV_CODEC_ID_SVQ3 = 23

AV_CODEC_ID_DVVIDEO = 24

AV_CODEC_ID_HUFFYUV = 25

AV_CODEC_ID_CYUV = 26

AV_CODEC_ID_H264 = 27

AV_CODEC_ID_INDEO3 = 28

AV_CODEC_ID_VP3 = 29

AV_CODEC_ID_THEORA = 30

AV_CODEC_ID_ASV1 = 31

AV_CODEC_ID_ASV2 = 32

AV_CODEC_ID_FFV1 = 33

AV_CODEC_ID_4XM = 34

AV_CODEC_ID_VCR1 = 35

AV_CODEC_ID_CLJR = 36

AV_CODEC_ID_MDEC = 37

AV_CODEC_ID_ROQ = 38

AV_CODEC_ID_INTERPLAY_VIDEO = 39

AV_CODEC_ID_XAN_WC3 = 40

AV_CODEC_ID_XAN_WC4 = 41

AV_CODEC_ID_RPZA = 42

AV_CODEC_ID_CINEPAK = 43

AV_CODEC_ID_WS_VQA = 44

AV_CODEC_ID_MSRLE = 45

AV_CODEC_ID_MSVIDEO1 = 46

AV_CODEC_ID_IDCIN = 47

AV_CODEC_ID_8BPS = 48

AV_CODEC_ID_SMC = 49

AV_CODEC_ID_FLIC = 50

AV_CODEC_ID_TRUEMOTION1 = 51

AV_CODEC_ID_VMDVIDEO = 52

AV_CODEC_ID_MSZH = 53

AV_CODEC_ID_ZLIB = 54

AV_CODEC_ID_QTRLE = 55

AV_CODEC_ID_TSCC = 56

AV_CODEC_ID_ULTI = 57

AV_CODEC_ID_QDRAW = 58

AV_CODEC_ID_VIXL = 59

AV_CODEC_ID_QPEG = 60

AV_CODEC_ID_PNG = 61

AV_CODEC_ID_PPM = 62

AV_CODEC_ID_PBM = 63

AV_CODEC_ID_PGM = 64

AV_CODEC_ID_PGMYUV = 65

AV_CODEC_ID_PAM = 66

AV_CODEC_ID_FFVHUFF = 67

AV_CODEC_ID_RV30 = 68

AV_CODEC_ID_RV40 = 69

AV_CODEC_ID_VC1 = 70

AV_CODEC_ID_WMV3 = 71

AV_CODEC_ID_LOCO = 72

AV_CODEC_ID_WNV1 = 73

AV_CODEC_ID_AASC = 74

AV_CODEC_ID_INDEO2 = 75

AV_CODEC_ID_FRAPS = 76

AV_CODEC_ID_TRUEMOTION2 = 77

AV_CODEC_ID_BMP = 78

AV_CODEC_ID_CSCD = 79

AV_CODEC_ID_MMVIDEO = 80

AV_CODEC_ID_ZMBV = 81

AV_CODEC_ID_AVS = 82

AV_CODEC_ID_SMACKVIDEO = 83

AV_CODEC_ID_NUV = 84

AV_CODEC_ID_KMVC = 85

AV_CODEC_ID_FLASHSV = 86

AV_CODEC_ID_CAVS = 87

AV_CODEC_ID_JPEG2000 = 88

AV_CODEC_ID_VMNC = 89

AV_CODEC_ID_VP5 = 90

AV_CODEC_ID_VP6 = 91

AV_CODEC_ID_VP6F = 92

AV_CODEC_ID_TARGA = 93

AV_CODEC_ID_DSICINVIDEO = 94

AV_CODEC_ID_TIERTEXSEQVIDEO = 95

AV_CODEC_ID_TIFF = 96

AV_CODEC_ID_GIF = 97

AV_CODEC_ID_DXA = 98

AV_CODEC_ID_DNXHD = 99

AV_CODEC_ID_THP = 100

AV_CODEC_ID_SGI = 101

AV_CODEC_ID_C93 = 102

AV_CODEC_ID_BETHSOFTVID = 103

AV_CODEC_ID_PTX = 104

AV_CODEC_ID_TXD = 105

AV_CODEC_ID_VP6A = 106

AV_CODEC_ID_AMV = 107

AV_CODEC_ID_VB = 108

AV_CODEC_ID_PCX = 109

AV_CODEC_ID_SUNRAST = 110

AV_CODEC_ID_INDEO4 = 111

AV_CODEC_ID_INDEO5 = 112

AV_CODEC_ID_MIMIC = 113

AV_CODEC_ID_RL2 = 114

AV_CODEC_ID_ESCAPE124 = 115

AV_CODEC_ID_DIRAC = 116

AV_CODEC_ID_BFI = 117

AV_CODEC_ID_CMV = 118

AV_CODEC_ID_MOTIONPIXELS = 119

AV_CODEC_ID_TGV = 120

AV_CODEC_ID_TGQ = 121

AV_CODEC_ID_TQI = 122

AV_CODEC_ID_AURA = 123

AV_CODEC_ID_AURA2 = 124

AV_CODEC_ID_V210X = 125

AV_CODEC_ID_TMV = 126

AV_CODEC_ID_V210 = 127

AV_CODEC_ID_DPX = 128

AV_CODEC_ID_MAD = 129

AV_CODEC_ID_FRWU = 130

AV_CODEC_ID_FLASHSV2 = 131

AV_CODEC_ID_CDGRAPHICS = 132

AV_CODEC_ID_R210 = 133

AV_CODEC_ID_ANM = 134

AV_CODEC_ID_BINKVIDEO = 135

AV_CODEC_ID_IFF_ILBM = 136

AV_CODEC_ID_KGV1 = 137

AV_CODEC_ID_YOP = 138

AV_CODEC_ID_VP8 = 139

AV_CODEC_ID_PICTOR = 140

AV_CODEC_ID_ANSI = 141

AV_CODEC_ID_A64_MULTI = 142

AV_CODEC_ID_A64_MULTI5 = 143

AV_CODEC_ID_R10K = 144

AV_CODEC_ID_MXPEG = 145

AV_CODEC_ID_LAGARITH = 146

AV_CODEC_ID_PRORES = 147

AV_CODEC_ID_JV = 148

AV_CODEC_ID_DFA = 149

AV_CODEC_ID_WMV3IMAGE = 150

AV_CODEC_ID_VC1IMAGE = 151

AV_CODEC_ID_UTVIDEO = 152

AV_CODEC_ID_BMV_VIDEO = 153

AV_CODEC_ID_VBLE = 154

AV_CODEC_ID_DXTORY = 155

AV_CODEC_ID_V410 = 156

AV_CODEC_ID_XWD = 157

AV_CODEC_ID_CDXL = 158

AV_CODEC_ID_XBM = 159

AV_CODEC_ID_ZEROCODEC = 160

AV_CODEC_ID_MSS1 = 161

AV_CODEC_ID_MSA1 = 162

AV_CODEC_ID_TSCC2 = 163

AV_CODEC_ID_MTS2 = 164

AV_CODEC_ID_CLLC = 165

AV_CODEC_ID_MSS2 = 166

AV_CODEC_ID_VP9 = 167

AV_CODEC_ID_AIC = 168

AV_CODEC_ID_ESCAPE130 = 169

AV_CODEC_ID_G2M = 170

AV_CODEC_ID_WEBP = 171

AV_CODEC_ID_HNM4_VIDEO = 172

AV_CODEC_ID_HEVC = 173

AV_CODEC_ID_FIC = 174

AV_CODEC_ID_ALIAS_PIX = 175

AV_CODEC_ID_BRENDER_PIX = 176

AV_CODEC_ID_PAF_VIDEO = 177

AV_CODEC_ID_EXR = 178

AV_CODEC_ID_VP7 = 179

AV_CODEC_ID_SANM = 180

AV_CODEC_ID_SGIRLE = 181

AV_CODEC_ID_MVC1 = 182

AV_CODEC_ID_MVC2 = 183

AV_CODEC_ID_HQX = 184

AV_CODEC_ID_TDSC = 185

AV_CODEC_ID_HQ_HQA = 186

AV_CODEC_ID_HAP = 187

AV_CODEC_ID_DDS = 188

AV_CODEC_ID_DXV = 189

AV_CODEC_ID_SCREENPRESSO = 190

AV_CODEC_ID_RSCC = 191

AV_CODEC_ID_AVS2 = 192

AV_CODEC_ID_PGX = 193

AV_CODEC_ID_AVS3 = 194

AV_CODEC_ID_MSP2 = 195

AV_CODEC_ID_VVC = 196

AV_CODEC_ID_Y41P = 197

AV_CODEC_ID_AVRP = 198

AV_CODEC_ID_012V = 199

AV_CODEC_ID_AVUI = 200

AV_CODEC_ID_TARGA_Y216 = 201

AV_CODEC_ID_V308 = 202

AV_CODEC_ID_V408 = 203

AV_CODEC_ID_YUV4 = 204

AV_CODEC_ID_AVRN = 205

AV_CODEC_ID_CPIA = 206

AV_CODEC_ID_XFACE = 207

AV_CODEC_ID_SNOW = 208

AV_CODEC_ID_SMVJPEG = 209

AV_CODEC_ID_APNG = 210

AV_CODEC_ID_DAALA = 211

AV_CODEC_ID_CFHD = 212

AV_CODEC_ID_TRUEMOTION2RT = 213

AV_CODEC_ID_M101 = 214

AV_CODEC_ID_MAGICYUV = 215

AV_CODEC_ID_SHEERVIDEO = 216

AV_CODEC_ID_YLC = 217

AV_CODEC_ID_PSD = 218

AV_CODEC_ID_PIXLET = 219

AV_CODEC_ID_SPEEDHQ = 220

AV_CODEC_ID_FMVC = 221

AV_CODEC_ID_SCPR = 222

AV_CODEC_ID_CLEARVIDEO = 223

AV_CODEC_ID_XPM = 224

AV_CODEC_ID_AV1 = 225

AV_CODEC_ID_BITPACKED = 226

AV_CODEC_ID_MSCC = 227

AV_CODEC_ID_SRGC = 228

AV_CODEC_ID_SVG = 229

AV_CODEC_ID_GDV = 230

AV_CODEC_ID_FITS = 231

AV_CODEC_ID_IMM4 = 232

AV_CODEC_ID_PROSUMER = 233

AV_CODEC_ID_MWSC = 234

AV_CODEC_ID_WCMV = 235

AV_CODEC_ID_RASC = 236

AV_CODEC_ID_HYMT = 237

AV_CODEC_ID_ARBC = 238

AV_CODEC_ID_AGM = 239

AV_CODEC_ID_LSCR = 240

AV_CODEC_ID_VP4 = 241

AV_CODEC_ID_IMM5 = 242

AV_CODEC_ID_MVDV = 243

AV_CODEC_ID_MVHA = 244

AV_CODEC_ID_CDTOONS = 245

AV_CODEC_ID_MV30 = 246

AV_CODEC_ID_NOTCHLC = 247

AV_CODEC_ID_PFM = 248

AV_CODEC_ID_MOBICLIP = 249

AV_CODEC_ID_PHOTOCD = 250

AV_CODEC_ID_IPU = 251

AV_CODEC_ID_ARGO = 252

AV_CODEC_ID_CRI = 253

AV_CODEC_ID_SIMBIOSIS_IMX = 254

AV_CODEC_ID_SGA_VIDEO = 255

AV_CODEC_ID_GEM = 256

AV_CODEC_ID_VBN = 257

AV_CODEC_ID_JPEGXL = 258

AV_CODEC_ID_QOI = 259

AV_CODEC_ID_PHM = 260

AV_CODEC_ID_RADIANCE_HDR = 261

AV_CODEC_ID_WBMP = 262

AV_CODEC_ID_MEDIA100 = 263

AV_CODEC_ID_VQC = 264

AV_CODEC_ID_PDV = 265

AV_CODEC_ID_EVC = 266

AV_CODEC_ID_RTV1 = 267

AV_CODEC_ID_VMIX = 268

AV_CODEC_ID_LEAD = 269

AV_CODEC_ID_PCM_S16LE = 65_536

AV_CODEC_ID_PCM_S16BE = 65_537

AV_CODEC_ID_PCM_U16LE = 65_538

AV_CODEC_ID_PCM_U16BE = 65_539

AV_CODEC_ID_PCM_S8 = 65_540

AV_CODEC_ID_PCM_U8 = 65_541

AV_CODEC_ID_PCM_MULAW = 65_542

AV_CODEC_ID_PCM_ALAW = 65_543

AV_CODEC_ID_PCM_S32LE = 65_544

AV_CODEC_ID_PCM_S32BE = 65_545

AV_CODEC_ID_PCM_U32LE = 65_546

AV_CODEC_ID_PCM_U32BE = 65_547

AV_CODEC_ID_PCM_S24LE = 65_548

AV_CODEC_ID_PCM_S24BE = 65_549

AV_CODEC_ID_PCM_U24LE = 65_550

AV_CODEC_ID_PCM_U24BE = 65_551

AV_CODEC_ID_PCM_S24DAUD = 65_552

AV_CODEC_ID_PCM_ZORK = 65_553

AV_CODEC_ID_PCM_S16LE_PLANAR = 65_554

AV_CODEC_ID_PCM_DVD = 65_555

AV_CODEC_ID_PCM_F32BE = 65_556

AV_CODEC_ID_PCM_F32LE = 65_557

AV_CODEC_ID_PCM_F64BE = 65_558

AV_CODEC_ID_PCM_F64LE = 65_559

AV_CODEC_ID_PCM_BLURAY = 65_560

AV_CODEC_ID_PCM_LXF = 65_561

AV_CODEC_ID_S302M = 65_562

AV_CODEC_ID_PCM_S8_PLANAR = 65_563

AV_CODEC_ID_PCM_S24LE_PLANAR = 65_564

AV_CODEC_ID_PCM_S32LE_PLANAR = 65_565

AV_CODEC_ID_PCM_S16BE_PLANAR = 65_566

AV_CODEC_ID_PCM_S64LE = 65_567

AV_CODEC_ID_PCM_S64BE = 65_568

AV_CODEC_ID_PCM_F16LE = 65_569

AV_CODEC_ID_PCM_F24LE = 65_570

AV_CODEC_ID_PCM_VIDC = 65_571

AV_CODEC_ID_PCM_SGA = 65_572

AV_CODEC_ID_ADPCM_IMA_QT = 69_632

AV_CODEC_ID_ADPCM_IMA_WAV = 69_633

AV_CODEC_ID_ADPCM_IMA_DK3 = 69_634

AV_CODEC_ID_ADPCM_IMA_DK4 = 69_635

AV_CODEC_ID_ADPCM_IMA_WS = 69_636

AV_CODEC_ID_ADPCM_IMA_SMJPEG = 69_637

AV_CODEC_ID_ADPCM_MS = 69_638

AV_CODEC_ID_ADPCM_4XM = 69_639

AV_CODEC_ID_ADPCM_XA = 69_640

AV_CODEC_ID_ADPCM_ADX = 69_641

AV_CODEC_ID_ADPCM_EA = 69_642

AV_CODEC_ID_ADPCM_G726 = 69_643

AV_CODEC_ID_ADPCM_CT = 69_644

AV_CODEC_ID_ADPCM_SWF = 69_645

AV_CODEC_ID_ADPCM_YAMAHA = 69_646

AV_CODEC_ID_ADPCM_SBPRO_4 = 69_647

AV_CODEC_ID_ADPCM_SBPRO_3 = 69_648

AV_CODEC_ID_ADPCM_SBPRO_2 = 69_649

AV_CODEC_ID_ADPCM_THP = 69_650

AV_CODEC_ID_ADPCM_IMA_AMV = 69_651

AV_CODEC_ID_ADPCM_EA_R1 = 69_652

AV_CODEC_ID_ADPCM_EA_R3 = 69_653

AV_CODEC_ID_ADPCM_EA_R2 = 69_654

AV_CODEC_ID_ADPCM_IMA_EA_SEAD = 69_655

AV_CODEC_ID_ADPCM_IMA_EA_EACS = 69_656

AV_CODEC_ID_ADPCM_EA_XAS = 69_657

AV_CODEC_ID_ADPCM_EA_MAXIS_XA = 69_658

AV_CODEC_ID_ADPCM_IMA_ISS = 69_659

AV_CODEC_ID_ADPCM_G722 = 69_660

AV_CODEC_ID_ADPCM_IMA_APC = 69_661

AV_CODEC_ID_ADPCM_VIMA = 69_662

AV_CODEC_ID_ADPCM_AFC = 69_663

AV_CODEC_ID_ADPCM_IMA_OKI = 69_664

AV_CODEC_ID_ADPCM_DTK = 69_665

AV_CODEC_ID_ADPCM_IMA_RAD = 69_666

AV_CODEC_ID_ADPCM_G726LE = 69_667

AV_CODEC_ID_ADPCM_THP_LE = 69_668

AV_CODEC_ID_ADPCM_PSX = 69_669

AV_CODEC_ID_ADPCM_AICA = 69_670

AV_CODEC_ID_ADPCM_IMA_DAT4 = 69_671

AV_CODEC_ID_ADPCM_MTAF = 69_672

AV_CODEC_ID_ADPCM_AGM = 69_673

AV_CODEC_ID_ADPCM_ARGO = 69_674

AV_CODEC_ID_ADPCM_IMA_SSI = 69_675

AV_CODEC_ID_ADPCM_ZORK = 69_676

AV_CODEC_ID_ADPCM_IMA_APM = 69_677

AV_CODEC_ID_ADPCM_IMA_ALP = 69_678

AV_CODEC_ID_ADPCM_IMA_MTF = 69_679

AV_CODEC_ID_ADPCM_IMA_CUNNING = 69_680

AV_CODEC_ID_ADPCM_IMA_MOFLEX = 69_681

AV_CODEC_ID_ADPCM_IMA_ACORN = 69_682

AV_CODEC_ID_ADPCM_XMD = 69_683

AV_CODEC_ID_AMR_NB = 73_728

AV_CODEC_ID_AMR_WB = 73_729

AV_CODEC_ID_RA_144 = 77_824

AV_CODEC_ID_RA_288 = 77_825

AV_CODEC_ID_ROQ_DPCM = 81_920

AV_CODEC_ID_INTERPLAY_DPCM = 81_921

AV_CODEC_ID_XAN_DPCM = 81_922

AV_CODEC_ID_SOL_DPCM = 81_923

AV_CODEC_ID_SDX2_DPCM = 81_924

AV_CODEC_ID_GREMLIN_DPCM = 81_925

AV_CODEC_ID_DERF_DPCM = 81_926

AV_CODEC_ID_WADY_DPCM = 81_927

AV_CODEC_ID_CBD2_DPCM = 81_928

AV_CODEC_ID_MP2 = 86_016

AV_CODEC_ID_MP3 = 86_017

< preferred ID for decoding MPEG audio layer 1, 2 or 3

AV_CODEC_ID_AAC = 86_018

AV_CODEC_ID_AC3 = 86_019

AV_CODEC_ID_DTS = 86_020

AV_CODEC_ID_VORBIS = 86_021

AV_CODEC_ID_DVAUDIO = 86_022

AV_CODEC_ID_WMAV1 = 86_023

AV_CODEC_ID_WMAV2 = 86_024

AV_CODEC_ID_MACE3 = 86_025

AV_CODEC_ID_MACE6 = 86_026

AV_CODEC_ID_VMDAUDIO = 86_027

AV_CODEC_ID_FLAC = 86_028

AV_CODEC_ID_MP3ADU = 86_029

AV_CODEC_ID_MP3ON4 = 86_030

AV_CODEC_ID_SHORTEN = 86_031

AV_CODEC_ID_ALAC = 86_032

AV_CODEC_ID_WESTWOOD_SND1 = 86_033

AV_CODEC_ID_GSM = 86_034

< as in Berlin toast format

AV_CODEC_ID_QDM2 = 86_035

AV_CODEC_ID_COOK = 86_036

AV_CODEC_ID_TRUESPEECH = 86_037

AV_CODEC_ID_TTA = 86_038

AV_CODEC_ID_SMACKAUDIO = 86_039

AV_CODEC_ID_QCELP = 86_040

AV_CODEC_ID_WAVPACK = 86_041

AV_CODEC_ID_DSICINAUDIO = 86_042

AV_CODEC_ID_IMC = 86_043

AV_CODEC_ID_MUSEPACK7 = 86_044

AV_CODEC_ID_MLP = 86_045

AV_CODEC_ID_GSM_MS = 86_046

AV_CODEC_ID_ATRAC3 = 86_047

AV_CODEC_ID_APE = 86_048

AV_CODEC_ID_NELLYMOSER = 86_049

AV_CODEC_ID_MUSEPACK8 = 86_050

AV_CODEC_ID_SPEEX = 86_051

AV_CODEC_ID_WMAVOICE = 86_052

AV_CODEC_ID_WMAPRO = 86_053

AV_CODEC_ID_WMALOSSLESS = 86_054

AV_CODEC_ID_ATRAC3P = 86_055

AV_CODEC_ID_EAC3 = 86_056

AV_CODEC_ID_SIPR = 86_057

AV_CODEC_ID_MP1 = 86_058

AV_CODEC_ID_TWINVQ = 86_059

AV_CODEC_ID_TRUEHD = 86_060

AV_CODEC_ID_MP4ALS = 86_061

AV_CODEC_ID_ATRAC1 = 86_062

AV_CODEC_ID_BINKAUDIO_RDFT = 86_063

AV_CODEC_ID_BINKAUDIO_DCT = 86_064

AV_CODEC_ID_AAC_LATM = 86_065

AV_CODEC_ID_QDMC = 86_066

AV_CODEC_ID_CELT = 86_067

AV_CODEC_ID_G723_1 = 86_068

AV_CODEC_ID_G729 = 86_069

AV_CODEC_ID_8SVX_EXP = 86_070

AV_CODEC_ID_8SVX_FIB = 86_071

AV_CODEC_ID_BMV_AUDIO = 86_072

AV_CODEC_ID_RALF = 86_073

AV_CODEC_ID_IAC = 86_074

AV_CODEC_ID_ILBC = 86_075

AV_CODEC_ID_OPUS = 86_076

AV_CODEC_ID_COMFORT_NOISE = 86_077

AV_CODEC_ID_TAK = 86_078

AV_CODEC_ID_METASOUND = 86_079

AV_CODEC_ID_PAF_AUDIO = 86_080

AV_CODEC_ID_ON2AVC = 86_081

AV_CODEC_ID_DSS_SP = 86_082

AV_CODEC_ID_CODEC2 = 86_083

AV_CODEC_ID_FFWAVESYNTH = 86_084

AV_CODEC_ID_SONIC = 86_085

AV_CODEC_ID_SONIC_LS = 86_086

AV_CODEC_ID_EVRC = 86_087

AV_CODEC_ID_SMV = 86_088

AV_CODEC_ID_DSD_LSBF = 86_089

AV_CODEC_ID_DSD_MSBF = 86_090

AV_CODEC_ID_DSD_LSBF_PLANAR = 86_091

AV_CODEC_ID_DSD_MSBF_PLANAR = 86_092

AV_CODEC_ID_4GV = 86_093

AV_CODEC_ID_INTERPLAY_ACM = 86_094

AV_CODEC_ID_XMA1 = 86_095

AV_CODEC_ID_XMA2 = 86_096

AV_CODEC_ID_DST = 86_097

AV_CODEC_ID_ATRAC3AL = 86_098

AV_CODEC_ID_ATRAC3PAL = 86_099

AV_CODEC_ID_DOLBY_E = 86_100

AV_CODEC_ID_APTX = 86_101

AV_CODEC_ID_APTX_HD = 86_102

AV_CODEC_ID_SBC = 86_103

AV_CODEC_ID_ATRAC9 = 86_104

AV_CODEC_ID_HCOM = 86_105

AV_CODEC_ID_ACELP_KELVIN = 86_106

AV_CODEC_ID_MPEGH_3D_AUDIO = 86_107

AV_CODEC_ID_SIREN = 86_108

AV_CODEC_ID_HCA = 86_109

AV_CODEC_ID_FASTAUDIO = 86_110

AV_CODEC_ID_MSNSIREN = 86_111

AV_CODEC_ID_DFPWM = 86_112

AV_CODEC_ID_BONK = 86_113

AV_CODEC_ID_MISC4 = 86_114

AV_CODEC_ID_APAC = 86_115

AV_CODEC_ID_FTR = 86_116

AV_CODEC_ID_WAVARC = 86_117

AV_CODEC_ID_RKA = 86_118

AV_CODEC_ID_AC4 = 86_119

AV_CODEC_ID_OSQ = 86_120

AV_CODEC_ID_QOA = 86_121

AV_CODEC_ID_LC3 = 86_122

AV_CODEC_ID_DVD_SUBTITLE = 94_208

AV_CODEC_ID_DVB_SUBTITLE = 94_209

AV_CODEC_ID_TEXT = 94_210

< raw UTF-8 text

AV_CODEC_ID_XSUB = 94_211

AV_CODEC_ID_SSA = 94_212

AV_CODEC_ID_MOV_TEXT = 94_213

AV_CODEC_ID_HDMV_PGS_SUBTITLE = 94_214

AV_CODEC_ID_DVB_TELETEXT = 94_215

AV_CODEC_ID_SRT = 94_216

AV_CODEC_ID_MICRODVD = 94_217

AV_CODEC_ID_EIA_608 = 94_218

AV_CODEC_ID_JACOSUB = 94_219

AV_CODEC_ID_SAMI = 94_220

AV_CODEC_ID_REALTEXT = 94_221

AV_CODEC_ID_STL = 94_222

AV_CODEC_ID_SUBVIEWER1 = 94_223

AV_CODEC_ID_SUBVIEWER = 94_224

AV_CODEC_ID_SUBRIP = 94_225

AV_CODEC_ID_WEBVTT = 94_226

AV_CODEC_ID_MPL2 = 94_227

AV_CODEC_ID_VPLAYER = 94_228

AV_CODEC_ID_PJS = 94_229

AV_CODEC_ID_ASS = 94_230

AV_CODEC_ID_HDMV_TEXT_SUBTITLE = 94_231

AV_CODEC_ID_TTML = 94_232

AV_CODEC_ID_ARIB_CAPTION = 94_233

AV_CODEC_ID_TTF = 98_304

AV_CODEC_ID_SCTE_35 = 98_305

< Contain timestamp estimated through PCR of program stream.

AV_CODEC_ID_EPG = 98_306

AV_CODEC_ID_BINTEXT = 98_307

AV_CODEC_ID_XBIN = 98_308

AV_CODEC_ID_IDF = 98_309

AV_CODEC_ID_OTF = 98_310

AV_CODEC_ID_SMPTE_KLV = 98_311

AV_CODEC_ID_DVD_NAV = 98_312

AV_CODEC_ID_TIMED_ID3 = 98_313

AV_CODEC_ID_BIN_DATA = 98_314

AV_CODEC_ID_SMPTE_2038 = 98_315

AV_CODEC_ID_LCEVC = 98_316

AV_CODEC_ID_PROBE = 102_400

< codec_id is not known (like AV_CODEC_ID_NONE) but lavf should attempt to identify it

AV_CODEC_ID_MPEG2TS = 131_072

< FAKE codec to indicate a raw MPEG-2 TS stream (only used by libavformat)

AV_CODEC_ID_MPEG4SYSTEMS = 131_073

< FAKE codec to indicate a MPEG-4 Systems stream (only used by libavformat)

AV_CODEC_ID_FFMETADATA = 135_168

< Dummy codec for streams containing only metadata information.

AV_CODEC_ID_WRAPPED_AVFRAME = 135_169

< Passthrough codec, AVFrames wrapped in AVPacket

AV_CODEC_ID_VNULL = 135_170

Dummy null video codec, useful mainly for development and debugging. Null encoder/decoder discard all input and never return any output.

AV_CODEC_ID_ANULL = 135_171

Dummy null audio codec, useful mainly for development and debugging. Null encoder/decoder discard all input and never return any output.

Implementations

impl AVCodecID

pub const AV_CODEC_ID_FIRST_AUDIO: AVCodecID = AVCodecID::AV_CODEC_ID_PCM_S16LE

impl AVCodecID

pub const AV_CODEC_ID_FIRST_SUBTITLE: AVCodecID = AVCodecID::AV_CODEC_ID_DVD_SUBTITLE

impl AVCodecID

pub const AV_CODEC_ID_FIRST_UNKNOWN: AVCodecID = AVCodecID::AV_CODEC_ID_TTF