ffmpeg_sys_next

Function av_opt_set

pub unsafe extern "C" fn av_opt_set(
    obj: *mut c_void,
    name: *const c_char,
    val: *const c_char,
    search_flags: c_int,
) -> c_int

@defgroup opt_set_funcs Option setting functions @{ Those functions set the field of obj with the given name to value.

@param[in] obj A struct whose first element is a pointer to an AVClass. @param[in] name the name of the field to set @param[in] val The value to set. In case of av_opt_set() if the field is not of a string type, then the given string is parsed. SI postfixes and some named scalars are supported. If the field is of a numeric type, it has to be a numeric or named scalar. Behavior with more than one scalar and +- infix operators is undefined. If the field is of a flags type, it has to be a sequence of numeric scalars or named flags separated by ‘+’ or ‘-’. Prefixing a flag with ‘+’ causes it to be set without affecting the other flags; similarly, ‘-’ unsets a flag. If the field is of a dictionary type, it has to be a ‘:’ separated list of key=value parameters. Values containing ‘:’ special characters must be escaped. @param search_flags flags passed to av_opt_find2. I.e. if AV_OPT_SEARCH_CHILDREN is passed here, then the option may be set on a child of obj.

@return 0 if the value has been set, or an AVERROR code in case of error: AVERROR_OPTION_NOT_FOUND if no matching option exists AVERROR(ERANGE) if the value is out of range AVERROR(EINVAL) if the value is not valid