ffmpeg_sys_next

Function swr_convert

pub unsafe extern "C" fn swr_convert(
    s: *mut SwrContext,
    out: *const *mut u8,
    out_count: c_int,
    in_: *const *const u8,
    in_count: c_int,
) -> c_int

Convert audio.

in and in_count can be set to 0 to flush the last few samples out at the end.

If more input is provided than output space, then the input will be buffered. You can avoid this buffering by using swr_get_out_samples() to retrieve an upper bound on the required number of output samples for the given number of input samples. Conversion will run directly without copying whenever possible.

@param s allocated Swr context, with parameters set @param out output buffers, only the first one need be set in case of packed audio @param out_count amount of space available for output in samples per channel @param in input buffers, only the first one need to be set in case of packed audio @param in_count number of input samples available in one channel

@return number of samples output per channel, negative value on error